Change Log - Added QueryExecuteXml - Changed Action from Row Object to DBAction, changed type from RowAction Integer - Removed duo function from QryExecute, It can now only send one request which can be differentiated by the number of parameters. This makes coding simpler for maintenance. - Added multiple field to QueryResult to give additional information about Queries to the Database.

+++ PMH/Repositories/ConnectionRepository.cs
@@ -94,7 +94,8 @@ namespace Plex.PMH.Repositories
 
         public void Remove(int i)
         {
-            ConnectionRepo.Remove(i);
+            if(ConnectionRepo.ContainsKey(i))
+                ConnectionRepo.Remove(i);
         }
 
         public void CheckIn(int ConnectionId)
